Create our own version of rp5 live so that the IRB prompt actually works under JRuby proper
Basically, it seems that ruby-processing always uses its own copy of JRuby
unless you pass --jruby to rp5. However, this copy is very old. So, I basically
patched Processing::Runner so that the copy of JRuby you're running will always
be used. That means you'll have to be running JRuby to use this, for right now.